+++ tests/model_serving/model_server/conftest.py
@@ -10,9 +10,11 @@ from kubernetes.dynamic import DynamicClient
 from kubernetes.dynamic.exceptions import ResourceNotFoundError
 from ocp_resources.cluster_service_version import ClusterServiceVersion
 from ocp_resources.config_map import ConfigMap
+from ocp_resources.daemonset import DaemonSet
 from ocp_resources.data_science_cluster import DataScienceCluster
 from ocp_resources.inference_service import InferenceService
 from ocp_resources.namespace import Namespace
+from ocp_resources.node import Node
 from ocp_resources.persistent_volume_claim import PersistentVolumeClaim
 from ocp_resources.resource import ResourceEditor
 from ocp_resources.secret import Secret
@@ -20,7 +22,15 @@ from ocp_resources.service_account import ServiceAccount
 from ocp_resources.serving_runtime import ServingRuntime
 from ocp_resources.storage_class import StorageClass
 from pytest_testconfig import config as py_config
+from timeout_sampler import TimeoutExpiredError, TimeoutSampler
 
+from tests.model_serving.model_server.kserve.model_cache.utils import (
+    LOCAL_MODEL_NODE_GROUP_NAME,
+    MODEL_CACHE_AGENT_DAEMONSET,
+    MODEL_CACHE_NODE_COUNT,
+    MODEL_CACHE_SIZE,
+    LocalModelNodeGroup,
+)
 from utilities.constants import (
     DscComponents,
     KServeDeploymentType,
@@ -36,9 +46,11 @@ from utilities.data_science_cluster_utils import (
 )
 from utilities.inference_utils import create_isvc
 from utilities.infra import (
+    get_data_science_cluster,
     is_disconnected_cluster,
     s3_endpoint_secret,
     update_configmap_data,
+    wait_for_dsc_status_ready,
 )
 from utilities.kueue_utils import (
     ClusterQueue,
@@ -376,6 +388,93 @@ def gpu_model_car_inference_service(
         volumes_mounts=deepcopy(x=PREDICT_RESOURCES["volume_mounts"]),
     ) as isvc:
         yield isvc
+
+
+# Model Cache Fixtures
+@pytest.fixture(scope="session")
+def model_cache_infra_ready(
+    admin_client: DynamicClient,
+) -> Generator[DataScienceCluster, Any, Any]:
+    """Enable ``kserve.modelCache`` in the DSC and wait for the agent DaemonSet.
+
+    Patches the DSC to set ``modelCache.managementState: Managed`` with a
+    ``cacheSize`` and two worker ``nodeNames``.  On teardown the
+    ``ResourceEditor`` restores the original DSC spec automatically.
+
+    Shared by both ``InferenceService`` (``kserve/model_cache``) and
+    ``LLMInferenceService`` (``llmd``) local model cache tests, since the
+    ``LocalModelNamespaceCache`` controller and its DSC toggle are common
+    infrastructure regardless of the workload CRD.
+    """
+    dsc = get_data_science_cluster(client=admin_client)
+    applications_namespace: str = py_config["applications_namespace"]
+
+    already_labeled = sorted(
+        [node.name for node in Node.get(client=admin_client, label_selector="kserve/localmodel=worker")],
+    )
+    all_workers = sorted(
+        [node.name for node in Node.get(client=admin_client, label_selector="node-role.kubernetes.io/worker")],
+    )
+
+    if len(already_labeled) >= MODEL_CACHE_NODE_COUNT:
+        selected_nodes = already_labeled[:MODEL_CACHE_NODE_COUNT]
+    elif len(all_workers) >= MODEL_CACHE_NODE_COUNT:
+        selected_nodes = all_workers[:MODEL_CACHE_NODE_COUNT]
+    else:
+        pytest.fail(f"Need at least {MODEL_CACHE_NODE_COUNT} worker nodes for model cache; found {len(all_workers)}")
+
+    with ResourceEditor(
+        patches={
+            dsc: {
+                "spec": {
+                    "components": {
+                        "kserve": {
+                            "modelCache": {
+                                "managementState": "Managed",
+                                "cacheSize": MODEL_CACHE_SIZE,
+                                "nodeNames": selected_nodes,
+                            }
+                        }
+                    }
+                }
+            }
+        }
+    ):
+        wait_for_dsc_status_ready(dsc_resource=dsc)
+
+        try:
+            for sample in TimeoutSampler(
+                wait_timeout=300,
+                sleep=10,
+                func=lambda: LocalModelNodeGroup(client=admin_client, name=LOCAL_MODEL_NODE_GROUP_NAME).exists,
+            ):
+                if sample:
+                    break
+        except TimeoutExpiredError:
+            pytest.fail(
+                f"LocalModelNodeGroup '{LOCAL_MODEL_NODE_GROUP_NAME}' did not appear "
+                f"within {300}s after enabling modelCache in DSC"
+            )
+
+        agent = DaemonSet(
+            client=admin_client,
+            name=MODEL_CACHE_AGENT_DAEMONSET,
+            namespace=applications_namespace,
+        )
+        try:
+            for sample in TimeoutSampler(
+                wait_timeout=300,
+                sleep=10,
+                func=lambda: agent.exists,
+            ):
+                if sample:
+                    break
+        except TimeoutExpiredError:
+            pytest.fail(
+                f"DaemonSet '{MODEL_CACHE_AGENT_DAEMONSET}' did not appear in '{applications_namespace}' within {300}s"
+            )
+
+        yield dsc
 
 
 # Kueue Fixtures
